fix: Add linkding namespace to base kustomization
Removing the linkding namespace from the deployment and adding it to the
base kustomization instead. This is a better choice as the deployment
manifest and other manifests that are created within the base
kustomization will inherit the value. Its better than adding a
namespace to each manifest that is created there